requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.3.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
PyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.19
typing-extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
